test_earley_parser.py
# ruff: noqa: S101
from earley_parser import EarleyParser, Grammar, Rule


def state_in_chart(chart, position, lhs, rhs, dot, start):
    """Check if a state is in the chart at the given position."""
    for state in chart[position].values():
        # We check if the LHS matches and if the RHS starts with the expected symbols up to the dot position
        if (
            state.rule.lhs == lhs
            and state.rule.rhs[:dot] == rhs[:dot]  # Match RHS up to the dot position
            and state.dot == dot
            and state.start == start
        ):
            return True
    return False


def test_left_recursion():
    rules = [
        Rule("S", ["S", "a"]),
        Rule("S", ["a"]),
    ]
    grammar = Grammar(rules, "S")
    parser = EarleyParser(grammar)
    tokens = ["a", "a", "a"]

    parse_tree = parser.parse(tokens)
    assert parse_tree is not None

    final_position = len(tokens)
    assert state_in_chart(parser.chart, final_position, "S", ["S"], 1, 0)
    assert state_in_chart(parser.chart, 0, "S", ["S"], 0, 0)
    assert state_in_chart(parser.chart, 1, "S", ["a"], 1, 0)
    assert state_in_chart(parser.chart, 2, "S", ["S", "a"], 2, 0)


def test_right_recursion():
    rules = [
        Rule("S", ["a", "S"]),
        Rule("S", ["a"]),
    ]
    grammar = Grammar(rules, "S")
    parser = EarleyParser(grammar)
    tokens = ["a", "a", "a"]

    parse_tree = parser.parse(tokens)
    assert parse_tree is not None

    final_position = len(tokens)
    # Adjust the test to check for the correct state that reflects right recursion
    assert state_in_chart(parser.chart, final_position, "S", ["a", "S"], 2, 0)
    assert state_in_chart(parser.chart, final_position, "S", ["a", "S"], 2, 1)
    # assert state_in_chart(parser.chart, final_position, "S", ["a", "S"], 2, 2)


def test_nullable_nonterminals():
    rules = [
        Rule("S", ["A", "B"]),
        Rule("A", ["a"]),
        Rule("A", []),  # Nullable production
        Rule("B", ["b"]),
    ]
    grammar = Grammar(rules, "S")
    parser = EarleyParser(grammar)
    tokens = ["b"]

    parse_tree = parser.parse(tokens)
    assert parse_tree is not None

    assert "A" in parser.nullable_nonterminals
    # Adjust the test to check for nullable productions
    assert state_in_chart(parser.chart, 1, "S", ["A", "B"], 2, 0)


def test_ambiguity():
    rules = [
        Rule("E", ["E", "+", "E"]),
        Rule("E", ["E", "*", "E"]),
        Rule("E", ["(", "E", ")"]),
        Rule("E", ["id"]),
    ]
    grammar = Grammar(rules, "E")
    parser = EarleyParser(grammar)
    tokens = ["id", "+", "id", "*", "id"]

    parse_tree = parser.parse(tokens)
    assert parse_tree is not None

    assert state_in_chart(parser.chart, len(tokens), "E", ["E"], 1, 0)


def test_complex_expression():
    rules = [
        Rule("E", ["E", "+", "E"]),
        Rule("E", ["E", "*", "E"]),
        Rule("E", ["id"]),
    ]
    grammar = Grammar(rules, "E")
    parser = EarleyParser(grammar)
    tokens = ["id", "*", "id", "+", "id"]

    parse_tree = parser.parse(tokens)
    assert parse_tree is not None

    final_position = len(tokens)
    assert state_in_chart(parser.chart, final_position, "E", ["E"], 1, 0)


def test_empty_input():
    rules = [
        Rule("S", []),  # Nullable start symbol
    ]
    grammar = Grammar(rules, "S")
    parser = EarleyParser(grammar)
    tokens = []

    parse_tree = parser.parse(tokens)
    assert parse_tree is not None
    assert "S" in parser.nullable_nonterminals
    # assert state_in_chart(parser.chart, 0, "S", ["S"], 1, 0)
    assert state_in_chart(parser.chart, 0, "S", [], 0, 0)


def test_invalid_input():
    rules = [
        Rule("S", ["a"]),
    ]
    grammar = Grammar(rules, "S")
    parser = EarleyParser(grammar)
    tokens = ["b"]  # Token not in grammar

    parse_tree = parser.parse(tokens)
    assert parse_tree is None

    final_position = len(tokens)
    start_state_found = any(
        state.rule.lhs == "S" and state.is_complete() for state in parser.chart[final_position].values()
    )
    assert not start_state_found


def test_leo_optimization_right_recursion():
    rules = [
        Rule("S", ["a", "S"]),
        Rule("S", ["a"]),
    ]
    grammar = Grammar(rules, "S")
    tokens = ["a", "a", "a"]

    parser = EarleyParser(grammar)
    parse_tree = parser.parse(tokens)
    assert parse_tree is not None

    for i, memo in enumerate(parser.leo_memo):
        if memo:
            assert "S" in memo, f"Leo memoization missing for 'S' at position {i}"


def test_deep_right_recursion():
    rules = [
        Rule("S", ["a", "S"]),
        Rule("S", ["a"]),
    ]
    grammar = Grammar(rules, "S")
    seq_len = [1, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100]
    chart_sizes = []

    for i in seq_len:
        tokens = ["a"] * i
        parser = EarleyParser(grammar, verbose=False)
        parse_tree = parser.parse(tokens)
        assert parse_tree is not None

        max_chart_size = max(len(states) for states in parser.chart)
        chart_sizes.append(max_chart_size)

    states_per_token = {
        (c - chart_sizes[0]) / (e - seq_len[0]) for e, c in zip(seq_len[1:], chart_sizes[1:], strict=False)
    }
    assert len(states_per_token) == 1, "Chart size is too large, Leo's optimization might not be working"
